backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
duckdb==1.1.3
google-generativeai==0.8.3
pydantic==2.10.3
python-dotenv==1.0.1
pandas==2.2.3
